backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdfplumber
python-dotenv
Pillow
google-generativeai
ollama